backend/requirements.txt
fastapi
uvicorn
httpx
python-multipart
google-generativeai
playwright
beautifulsoup4
supabase
python-dotenv
